feat #72 Provider 계층 추가(캐시/DB 로직)

+++ src/main/java/com/solo/ptmatch/trainer/presentation/response/S2ClusterResponse.java
@@ -7,9 +7,9 @@ public record S2ClusterResponse(
         Long s2CellId,
         Double latitude,
         Double longitude,
-        Integer trainerCount) {
+        Long trainerCount) {
 
-    public static S2ClusterResponse fromCellId(Long cellId, Integer count) {
+    public static S2ClusterResponse fromCellId(Long cellId, Long count) {
         S2CellId s2CellId = new S2CellId(cellId);
         S2LatLng center = s2CellId.toLatLng();
         return new S2ClusterResponse(
